adding exists
