+++ mutations/README.md
@@ -1,0 +1,8 @@
+# Mutations
+
+A set of functions to help process any types of mutations
+
+
+## contains:
+
+vcf_to_df: transforms a vcf file into a dataframe file as best as it can
